feat(wyvern-plugin): More progress with `Wyvern` plugin, added predicate functions.
