+++ src/plugin/nft/README.md
@@ -1,0 +1,3 @@
+# NFT Plugin
+
+[TBC]
