+++ src/plugin/wyvern-protocol/README.md
@@ -1,0 +1,5 @@
+# Wyvern Protocol Plugin
+
+## Predicates
+
+The following [predicates](predicates/README.md) are shipped with the protocol.
