+++ src/plugin/wyvern-protocol/predicates/README.md
@@ -1,0 +1,10 @@
+# Wyvern Protocol Predicates
+
+The definition of the supported predicate functions shipped with the protocol.
+
+* TransferERC721Exact
+* TransferERC20Exact
+* SequenceAnyAfter
+* Split
+* TransferERC20ExactTo
+* SequenceExact
